Updated error in printing last digit

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -7,7 +7,7 @@
  *
  * Return: the value of the last digit
  */
-int print_last_digit(int n)
+int print_last_digit(long int n)
 {
 	if (n < 0)
 	{
